+++ src/voxel/tests.rs
@@ -1,9 +1,8 @@
+use self::value::Empty;
 use crate::dynamic_vdb;
 use crate::static_vdb;
 use crate::voxel::utils::box_indices;
 use crate::voxel::*;
-
-use self::value::Empty;
 
 type StaticTree = static_vdb!(Empty, 4, 3, 2);
 type DynamicTree = dynamic_vdb!(Empty, 4, 3, 2);
