+++ src/voxel/value/mod.rs
@@ -1,0 +1,4 @@
+mod empty;
+mod f32;
+
+pub use empty::Empty;
